Made it a proper cog
- Added __init__
- Made the commands start with "@command.command()"
- Made the cog name `Admin`
- Added ✅ at the start of every command
